Refactor color well gradient
