+++ packages/colorpickers/.size-snapshot.json
@@ -1,20 +1,20 @@
 {
   "index.cjs.js": {
-    "bundled": 87839,
-    "minified": 60994,
-    "gzipped": 12245
+    "bundled": 87910,
+    "minified": 61058,
+    "gzipped": 12246
   },
   "index.esm.js": {
-    "bundled": 82853,
-    "minified": 56560,
-    "gzipped": 12096,
+    "bundled": 82924,
+    "minified": 56624,
+    "gzipped": 12099,
     "treeshaked": {
       "rollup": {
-        "code": 46400,
+        "code": 46477,
         "import_statements": 812
       },
       "webpack": {
-        "code": 50805
+        "code": 50884
       }
     }
   }
