+++ packages/colorpickers/src/styled/index.ts
@@ -10,6 +10,7 @@ export * from './ColorPicker/StyledHue';
 export * from './ColorPicker/StyledAlpha';
 export * from './ColorPicker/StyledPreview';
 export * from './ColorPicker/StyledColorWell';
+export * from './ColorPicker/StyledColorWellGradient';
 export * from './ColorPicker/StyledColorWellWrapper';
 export * from './ColorPicker/StyledColorWellThumb';
 export * from './ColorPicker/StyledSliderGroup';
